Added more comments
Also moved properties from CustomWindow to CustomDialog and cleaned up OnWindowReorder method.
